Github issue 34380: Replaced "int" to "float".

+++ app/code/Magento/Checkout/Controller/Sidebar/UpdateItemQty.php
@@ -66,7 +66,7 @@ class UpdateItemQty extends Action
     public function execute()
     {
         $itemId = (int)$this->getRequest()->getParam('item_id');
-        $itemQty = (int)$this->getRequest()->getParam('item_qty') * 1;
+        $itemQty = (float)$this->getRequest()->getParam('item_qty') * 1;
         $itemQty = $this->quantityProcessor->prepareQuantity($itemQty);
 
         try {
